modules/requirements.txt
astral
geopy
gpiozero
pynmea2
pyserial
pytz
Requests
timezonefinder
